Try add username to branch name to debug pull request validation error
From https://developer.github.com/v3/pulls/#create-a-pull-request: the head parameter: "The name of the branch where your changes are implemented. For cross-repository pull requests in the same network, namespace head with a user like this: username:branch."

This isn't a cross-repository pull request, so this shouldn't matter, but I want to see if I get a different error.